Change the diffusion codes
